- created initial ImageServiceInterface and its implementation for this project -> ImageServiceCloudinaryImpl.java
